django_microservices/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
psycopg[binary]
kafka-python==2.0.2
python-decouple==3.8
django-cors-headers==4.4.0
gunicorn==21.2.0
whitenoise==6.6.0
pillow==10.4.0
requests
djangorestframework-simplejwt
PyJWT
py_eureka_client
minio
django-storages
boto3
dotenv